configuracion de base de datos 11-06-2015

+++ src/Conexiones_SQL/GuardarSQL.java
@@ -21,6 +21,7 @@ import javax.swing.JOptionPane;
 
 import Obj_Administracion_del_Sistema.Obj_Asistencia_Y_Puntualidad;
 import Obj_Administracion_del_Sistema.Obj_Configuracion_Base_de_Datos;
+import Obj_Administracion_del_Sistema.Obj_Configuracion_Base_de_Datos_2;
 import Obj_Administracion_del_Sistema.Obj_Usuario;
 import Obj_Auditoria.Obj_Actividades_Por_Proyecto;
 import Obj_Auditoria.Obj_Actividades_Relacionadas;
@@ -1577,6 +1578,55 @@ public class GuardarSQL {
 				
 			}else{
 				File folder = new File(System.getProperty("user.dir")+"\\Config");
+				folder.mkdirs();
+				archivo.createNewFile();
+				bufferedWriter = new BufferedWriter (new FileWriter(nomArchivo));
+				
+				bufferedWriter.write(config.getDireccionIPV4()+    		"\n");
+				bufferedWriter.write(config.getNombreBD()+      		"\n");
+				bufferedWriter.write(config.getUsuario()+ 	    		"\n");
+				bufferedWriter.write(config.getContrasena()+       		"\n");
+				
+			}
+			
+		}
+		catch(FileNotFoundException ex)
+		{
+			ex.printStackTrace();
+		}catch(IOException ex)
+		{
+			ex.printStackTrace();
+		}finally
+		{
+			try
+			{
+				if(bufferedWriter!=null)
+				{
+					bufferedWriter.flush();
+					bufferedWriter.close();
+				}
+			}catch(IOException ex)
+			{
+				ex.printStackTrace();
+			}
+		}return true;
+	}
+	
+	public boolean Guardar_ConfigBD_2(Obj_Configuracion_Base_de_Datos_2 config){
+		BufferedWriter bufferedWriter = null;
+		String nomArchivo = System.getProperty("user.dir")+"\\Config\\config2";
+		try{
+			File archivo = new File(nomArchivo);
+			if(archivo.exists()){
+				bufferedWriter = new BufferedWriter (new FileWriter(nomArchivo));
+				
+				bufferedWriter.write(config.getDireccionIPV4()+    		"\n");
+				bufferedWriter.write(config.getNombreBD()+      		"\n");
+				bufferedWriter.write(config.getUsuario()+ 	    		"\n");
+				bufferedWriter.write(config.getContrasena()+       		"\n");
+				
+			}else{
+				File folder = new File(System.getProperty("user.dir")+"\\Config2");
 				folder.mkdirs();
 				archivo.createNewFile();
 				bufferedWriter = new BufferedWriter (new FileWriter(nomArchivo));
